Support Signature as an inner element with EnvelopedSignatureWriter
* Use InnerWriter to write canonical bytes (without signature).
* Use InternalWriter to write non-canonical xml (including signature).
* When WriteSignature() method is called, write only a placeholder
signature element using InternalWriter.
* After writing the last (end) element, create XmlTokenStream out of the
InternalWriter's stream and write that XmlTokenStream into the original
writer. When writing the XmlTokenStream, replace signature [placeholder]
element with the real signature (using DSigSerializer). This behavior is
only if WriteSignature method was invoked explicitly. Otherwise, the
default behavior is to generate the signature at the end and to write it
directly to the original stream.
* Add RoundTripWsMetadata test with a signature placed at different
positions.
* Add RountTripSamlp test to showcase writing an xml with multiple
signatures and later reading it at with an EnvelopedSignatureReader
and validating both inner and outer signature.